moved all matrix computations from trainbust
They're all in internal functions now, so should be way easier to
write derivatives for. The larger program is now just the logic
(which is ugly...).

Furthermore, BUG!!! pushing bulges up from the bottom, the shifts
do what we want (eigenvalue shifts produce zeros). When pushing
stuff towards the bottom, this is no longer true. damn.