fixing trainBust errors
I extracted the spikes, and the derivatives were exploding
to infinity.
I have so far discovered (and fixed) the following issues causing this
1) derivatives need to be updated before the vector itself.
2) can't treat the householder transformation as 2 different steps and
apply product rule. That's a convenient shorthand for (I-vv')H(I-vv'),
which yields a different derivative

Doing this introduces the bug that the v calculated doesn't actually
yield a vv' of same size as H, we're just carful applying it correctly.
I have failed to maintain such caution in the derivative, which is
currently broken.